requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymupdf4llm==0.0.17
python-multipart==0.0.6
pydantic==2.4.2
openai==1.3.5
python-dotenv==1.0.0
aiofiles==23.2.0
pypdf2==3.0.1
reportlab==4.0.4
